Allow instructors to pre-register students
