+++ views/admin/admin.html
@@ -93,6 +93,19 @@
 
             </div>
         </form>
+        <div class="col-md-6">
+            <h3 style="text-align: center">Register Students</h3>
+            <p>Upload a csv file with the following format to automatically register your students.
+            Each row must have six elements in the order shown below.</p>
+            <pre>username, email, first_name, last_name, password, course</pre>
+        <form method="POST" action="/{{=request.application}}/admin/enroll_students" enctype="multipart/form-data">
+            <div class="form-group">
+                <label for="studentuploads">Choose a CSV file with student information</label>
+                <input type="file" name="students">
+            </div>
+            <input type="submit" name="Send" value="Register Students">
+        </form>
+        </div>
     </div>
 
     <div id="sections" class="tab-pane fade">
